Fixed settings/performance/storage - bar info showing incorrect statistics

+++ settingsMenu.c
@@ -949,9 +949,9 @@ void storageMenu()
 	if (!performancebg || !highlight)
 		debugDisplay();
 	
-	int usedStorage = (storageGetTotalSize() - storageGetFreeSize());
-	int totalStorage = storageGetTotalSize();
-	int fill = ((usedStorage/totalStorage) * 404);
+	double usedStorage = (storageGetTotalSize() - storageGetFreeSize());
+	double totalStorage = storageGetTotalSize();
+	double fill = ((usedStorage / totalStorage) * 404.0);
 	
 	while (!osl_quit)
 	{
@@ -970,18 +970,18 @@ void storageMenu()
 		else
 			oslIntraFontSetStyle(Roboto, 0.9, LITEGRAY, 0, INTRAFONT_ALIGN_LEFT);
 		
-		oslDrawStringf(20, 80, "%.2d MB", usedStorage); 
+		oslDrawStringf(20, 80, "%.2f MB", usedStorage); 
 		
 		if (DARK == 0)
 			oslIntraFontSetStyle(Roboto, fontSize, RGBA(fontColor.r, fontColor.g, fontColor.b, 255), 0, INTRAFONT_ALIGN_LEFT);
 		else
 			oslIntraFontSetStyle(Roboto, fontSize, LITEGRAY, 0, INTRAFONT_ALIGN_LEFT);
 		
-		oslDrawStringf(20, 100, "Total used of %.2d MB", totalStorage); 
+		oslDrawStringf(20, 100, "Total used of %.2f MB", totalStorage); 
 		
 		
 		oslDrawStringf(20, 136, "Storage:"); 
-		oslDrawStringf(20, 152, "Total used of %.2d MB", totalStorage); 
+		oslDrawStringf(20, 152, "Total used of %.2f MB", totalStorage); 
 		
 		oslDrawFillRect(20, 172, 424, 176, RGB(206, 215, 219));
 		oslDrawFillRect(20, 172, fill, 176, RGB(r, g, b));
